+++ src/main/java/dk/easv/mymovies/GUI/Controller/AddEditMovieController.java
@@ -155,8 +155,8 @@ public class AddEditMovieController {
     private void saveMovie() {
         try {
             String name = tfdName.getText();
-            float imdbRating = Float.parseFloat(tfdIRating.getText().replace(',', '.'));
-            float yRating = Float.parseFloat(tfdPRating.getText().replace(',', '.'));
+            double imdbRating = Double.parseDouble(tfdIRating.getText().replace(',', '.'));
+            double yRating = Double.parseDouble(tfdPRating.getText().replace(',', '.'));
             String fileLink = lblMovieFileName.getText();
             String lastView = ""; // Set this to the appropriate value
             String posterLink = lblPosterFileName.getText();
